Redirect to author page if query exactly pauseid

+++ lib/MetaCPAN/Web/Controller/Search.pm
@@ -46,9 +46,26 @@ sub index : Path {
         )
     {
         my $module = $model->first($query)->recv;
-        $c->detach('/not_found') unless ($module);
-        $c->res->redirect("/pod/$module");
-        $c->detach;
+        if ( $query eq $module ) {
+            $c->res->redirect("/pod/$module");
+            $c->detach;
+        }
+        else {
+            my $author = $c->model('API::Author')->search($query)->recv;
+            if (   $author->{total} == 1
+                && $query eq $author->{results}->[0]->{pauseid} )
+            {
+                $c->res->redirect( '/author/' . uc($query) );
+                $c->detach;
+            }
+            elsif ($module) {
+                $c->res->redirect("/pod/$module");
+                $c->detach;
+            }
+            else {
+                $c->detach('/not_found') unless ($module);
+            }
+        }
     }
     else {
         my $user = $c->user_exists ? $c->user->id : undef;
